+++ Assets/Scripts/Egg.cs
@@ -253,7 +253,7 @@ public class Egg : PooledObject, IHitPoints
         }
     }
 
-    public IHitPoints.DamageReturn Damage(float damage, IHitPoints.HitType hitType)
+    public IHitPoints.DamageReturn Damage(int damage, IHitPoints.HitType hitType)
     {
         if (State.SPIN != m_state)
         {
